+++ mod/config.php
@@ -18,8 +18,8 @@
  * --------------------------------------------------------------------
  * Parameters of the ZnetDK 4 Mobile Home Menu module
  *
- * File version: 1.0
- * Last update: 05/30/2024
+ * File version: 1.1
+ * Last update: 07/04/2024
  */
 
 /**
@@ -35,24 +35,12 @@ define('MOD_Z4M_HOMEMENU_MAX_PANELS_PER_ROW', 4);
 define('MOD_Z4M_HOMEMENU_PANEL_ICON_WIDTH', '100px');
 
 /**
- * Sets the column definition for displaying menu panels
- * @var array Grid CSS classes to apply according to the number of panels to
- * display for each row.
- */
-define('MOD_Z4M_HOMEMENU_COLUMN_DEFINITION', [
-    1 => 'm12 l12',
-    2 => 'm6 l6',
-    3 => 'm6 l4',
-    4 => 'm6 l3'
-]);
-
-/**
  * Module version number
  * @var string Version
  */
-define('MOD_Z4M_HOMEMENU_VERSION_NUMBER','1.0');
+define('MOD_Z4M_HOMEMENU_VERSION_NUMBER','1.1');
 /**
  * Module version date
  * @var string Date in W3C format
  */
-define('MOD_Z4M_HOMEMENU_VERSION_DATE','2024-05-30');
+define('MOD_Z4M_HOMEMENU_VERSION_DATE','2024-07-04');