+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGE LOG: Home Menu (z4m_homemenu)
 
+## Version 1.1, 2024-07-04
+- CHANGE: the cells displaying the level 2 menu items have now the same height
+  on medium and large screens.
+
 ## Version 1.0, 2024-05-30
 First version.